Convert the project to typescript and improve typescript
This changes convert the project to typescript to address several
issues.

It also simplify the interface a bit by conveying to a callback pattern
of `request`, `token` and `callback` in the `isRevoked` and `secret`
functions.

Removes also the feature of changing the output for the decoded token,
`userProperty`, `requestProperty` and `resultProperty`. The middleware
puts the decoded token always on `req.auth`.

Changes everything to use promises.